requirements.txt
alembic==1.16.5
babel==2.17.0
bleach==6.3.0
blinker==1.9.0
boto3==1.40.21
botocore==1.40.21
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
croniter==6.0.0
Deprecated==1.2.18
Flask==3.1.2
flask-babel==4.0.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==7.1.0
reportlab==4.4.5
rich==13.9.4
rq==2.6.0
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
waitress==3.0.2
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
